+++ docs/changelog.rst
@@ -22,6 +22,7 @@ Fixed
 - Allow models to not have a spatial component (allows updating if region is None). (#1244)
 - CF compliant dimensions in netcdf files
 - `buffer` argument in `DataCatalog.get_rasterdataset` now is an integer expressed in resolution multiplicity instead of in meters.
+- `_parse_geom_bbox_buffer` has been renamed to `_parse_geom_bbox` and the buffering part has been removed.
 - Improved logging information for build and update functions. (#1237)
 
 Deprecated